Element R6 Class representing a UI Element

This is the abstract base class for all UI elements.

Value

top-level class name element's uuid

Details

An object that is meant to be the base type for all user-interface elements. The primary driver for having a base type for all UI elements is that every ui should have a unique id for shiny module dispatch, and basic descriptors that assist with robust logging.

Methods

Public methods

Element$new()

Element$class_name()

Element$identifier()

Element$clone()

Method new()

Creates a new instance of this R6 class.

Usage

Element$new()

Returns

A new Element object.

Method class_name()

Get accessor for (top-level) class name.

Usage

Element$class_name()

Method identifier()

Get accessor for object's uuid.

Usage

Element$identifier()

Method clone()

The objects of this class are cloneable with this method.

Usage

Element$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.