Element

An object that is meant to be the base type for all ui pages. Pages are meant to represent an entire page in the UI as serves as a container for controls.

References

Element

Super class

shiny.gentelella::Element -> Page

Methods

Public methods

Page$new()

Page$page()

Page$server()

Page$clone()

Inherited methods

shiny.gentelella::Element$class_name()

shiny.gentelella::Element$identifier()

Method new()

Create a new UI Page object.

Usage

Page$new()

Returns

A new Page object.

Method page()

Wrapper around fluid layout page.

Usage

Page$page()

Method server()

server-side wrapper.

Usage

Page$server(input, output, session)

Arguments

input

form input

output

render output

session

page session

Method clone()

The objects of this class are cloneable with this method.

Usage

Page$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.