Table

An object that is designed to encapsulate most of the standard plumbing in setting up a Datatable in the UI. The html template

References

Page

Super classes

shiny.gentelella::Element -> shiny.gentelella::Control -> PortalTable

Public fields

title

table title

export_btn

export btn

Methods

Public methods

Table$new()

Table$subtitle()

Table$ui()

Table$ui_extra()

Table$ui_footer()

Table$clone()

Inherited methods

shiny.gentelella::Element$class_name()

shiny.gentelella::Element$identifier()

Method new()

initialize an instance of a portal table.

Usage

Table$new(title)

Arguments

title

title of the table

Method subtitle()

function to customize the table subtitle.

Usage

Table$subtitle()

Method ui()

render the ui for the a portal styled data table.

Usage

Table$ui()

Method ui_extra()

pure virtual method for customizing the table UI.

Usage

Table$ui_extra()

Method ui_footer()

pure virtual method for customizing the table UI.

Usage

Table$ui_footer()

Method clone()

The objects of this class are cloneable with this method.

Usage

Table$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.